adding Stream and StreamCard
