+++ src/pages/components/StreamCard.jsx
@@ -1,0 +1,6 @@
+
+function StreamCard(){
+
+}
+
+export default StreamCard
